refactor(gate): tighten connector boundary
