+++ lib/channel_gate.mli
@@ -14,7 +14,7 @@
 
 (** Message arriving from an external channel consumer. *)
 type inbound_message = {
-  channel : Agent_identity.channel;
+  channel : string;
   channel_user_id : string;
   channel_user_name : string;
   channel_room_id : string;
