+++ sidecars/discord-bot/.env.example
@@ -2,10 +2,10 @@
 DISCORD_BOT_TOKEN=
 
 # masc-mcp server URL
-MASC_MCP_URL=http://localhost:8935
+GATE_BASE_URL=http://localhost:8935
 
 # masc-mcp Bearer token for /api/v1/gate/* endpoints
-MASC_API_TOKEN=
+GATE_API_TOKEN=
 
 # Channel-to-keeper mapping (JSON)
 # Keys: Discord channel snowflake IDs
