+++ sidecars/discord-bot/tests/test_masc_client.py
@@ -9,21 +9,21 @@ import pytest
 
 from src import config as config_module
 from src.config import BotConfig
-from src.masc_client import MascGateClient
+from src.gate_client import GateClient
 
 
 @pytest.fixture(autouse=True)
 def reset_config(monkeypatch: pytest.MonkeyPatch) -> Iterator[None]:
     monkeypatch.setenv("DISCORD_BOT_TOKEN", "test-token")
-    monkeypatch.setenv("MASC_API_TOKEN", "test-api-token")
-    monkeypatch.setenv("MASC_MCP_URL", "http://localhost:8935")
+    monkeypatch.setenv("GATE_API_TOKEN", "test-api-token")
+    monkeypatch.setenv("GATE_BASE_URL", "http://localhost:8935")
     config_module.reset_config_cache()
     yield
     config_module.reset_config_cache()
 
 
-def make_client(handler: httpx.MockTransport) -> MascGateClient:
-    client = MascGateClient()
+def make_client(handler: httpx.MockTransport) -> GateClient:
+    client = GateClient()
     client._client = httpx.AsyncClient(transport=handler, headers=client._headers)  # pyright: ignore[reportPrivateUsage]
     client._max_retries = 1  # pyright: ignore[reportPrivateUsage]
     client._breaker_failure_threshold = 2  # pyright: ignore[reportPrivateUsage]
@@ -130,7 +130,7 @@ def test_config_allows_zero_disable_values(monkeypatch: pytest.MonkeyPatch) -> N
 
     cfg = BotConfig(
         discord_bot_token="test-token",
-        masc_api_token="test-api-token",
+        gate_api_token="test-api-token",
         gate_max_retries=0,
         status_cache_ttl_sec=0,
         keeper_cache_ttl_sec=0,
